Update the state to hold ciphertext data for now
