+++ console/program/src/data/to_bits.rs
@@ -16,12 +16,14 @@
 
 use super::*;
 
-impl<N: Network> State<N> {
-    /// Returns the serial number of the record.
-    pub fn to_serial_number(&self, private_key: &PrivateKey<N>, randomizer: N::Scalar) -> Result<SerialNumber<N>> {
-        // Compute the commitment for the program state.
-        let commitment = self.to_commitment()?;
-        // Compute the serial number.
-        SerialNumber::<N>::prove(&private_key.sk_vrf(), commitment, randomizer)
+impl<N: Network, Private: Visibility<N>> ToBits for Data<N, Private> {
+    /// Returns this data as a list of **little-endian** bits.
+    fn to_bits_le(&self) -> Vec<bool> {
+        self.0.to_bits_le()
+    }
+
+    /// Returns this data as a list of **big-endian** bits.
+    fn to_bits_be(&self) -> Vec<bool> {
+        self.0.to_bits_be()
     }
 }
